requirements.txt
pyspark>=3.5.0
apache-airflow>=2.8.0
pandas-market-calendars>=4.0.0
yfinance>=0.2.0
pytest>=7.0.0
